Remove gradle from project, use Maven instead
There should be no real functional changes to the project, just
the build mechanism is now maven instead of Gradle.

The one important change to note is that Version.java is now
removed. lightstep-tracer-java and lightstep-tracer-android
should each define their own Version.java. It makes more sense
for those to each maintain a version now that the version
numbers will not always be in sync with common.